requirements.txt
django-filter
beautifulsoup4==4.8.1
Django==2.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-crispy-forms==1.8.0
django-getenv==1.3.2
Pillow==6.2.1
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
